moveonly: add mp/type-vector.h
